Update README.md
Hey team 
(cc: @MadMax1960),

I'm Fernando (fernhwb) on GitHub, and I was the producer for the initial 1.0.0 release back in early 2024. I set up the original README (still in use today), and created/managed to completion hundreds of Trello tasks that laid the groundwork for everything that followed. Unfortunately, my name got dropped from credits as the project grew, but I'd love to be re-added under a new "Founding/Producer" section or as a Major Contributor for Project Setup.

Happy to hop on Discord or provide commit history/Trello exports to verify, I really love this * a lot * and would love to still be listed.